Вносит правки в задание модуляция
