+++ js/util.js
@@ -31,59 +31,4 @@ function getRandomInteger(min, max) {
   return Math.floor(result);
 }
 
-function getRandomPhotoObjects({ photoGeneratorParameters }) {
-  const {
-    paramCount,
-    paramDescriptions,
-    paramRandomValue,
-    paramObjectsCount,
-    paramCommentsCount,
-    paramAvatarsCount,
-    paramSentencesCount,
-    paramLikesCount,
-    paramMessages,
-    paramNames
-  } = photoGeneratorParameters;
-
-  const objects = [];
-  const getUniqueCommentIdValue = getUniqueValue(getRandomInteger, paramRandomValue.MIN, paramRandomValue.MAX);
-  const getUniquePhotoIdValue = getUniqueValue(getRandomInteger, paramObjectsCount.MIN, paramObjectsCount.MAX);
-  const getUniqueUrlValue = getUniqueValue(getRandomInteger, paramObjectsCount.MIN, paramObjectsCount.MAX);
-
-
-  for (let i = 0; i < paramCount; i++) {
-    const comments = [];
-    const commentsCount = getRandomInteger(paramCommentsCount.MIN, paramCommentsCount.MAX);
-
-    for (let j = 0; j < commentsCount; j++) {
-      const commentId = getUniqueCommentIdValue();
-      const avatar = `img/avatar-${getRandomInteger(paramAvatarsCount.MIN, paramAvatarsCount.MAX)}.svg`;
-      const commentsList = getRandomArrayElements(paramMessages, getRandomInteger, paramSentencesCount.MAX);
-      const name = getRandomArrayElements(paramNames, getRandomInteger);
-
-      comments.push({
-        commentId,
-        avatar,
-        commentsList,
-        name
-      });
-    }
-
-    const photoId = getUniquePhotoIdValue();
-    const url = `photos/${getUniqueUrlValue()}.jpg`;
-    const description = paramDescriptions[getRandomInteger(0, paramDescriptions.length)];
-    const likes = getRandomInteger(paramLikesCount.MIN, paramLikesCount.MAX);
-
-    objects.push({
-      photoId,
-      url,
-      description,
-      likes,
-      comments
-    });
-  }
-
-  return objects;
-}
-
-export {getRandomPhotoObjects};
+export {getUniqueValue, getRandomInteger, getRandomArrayElements};
